KA-137 - Move reducers into corresponding folders
