+++ onboarding-task/.idea/codeStyleSettings.xml
@@ -17,7 +17,7 @@
           </value>
         </option>
         <option name="LINE_SEPARATOR" value="&#10;" />
-        <option name="RIGHT_MARGIN" value="100" />
+        <option name="RIGHT_MARGIN" value="150" />
         <option name="HTML_ATTRIBUTE_WRAP" value="4" />
         <option name="HTML_TEXT_WRAP" value="0" />
         <option name="HTML_KEEP_BLANK_LINES" value="4" />
