+++ onboarding-task/src/reducers/ListItemReducers/flagReducers/fetchingFlagReducer.ts
@@ -1,5 +1,5 @@
-import { IAction } from '../actionCreators/IAction';
-import { FETCH_ITEMS_FAILURE, FETCH_ITEMS_REQUEST, FETCH_ITEMS_SUCCESS } from '../constants/actionTypes';
+import { IAction } from '../../../actionCreators/IAction';
+import { FETCH_ITEMS_FAILURE, FETCH_ITEMS_REQUEST, FETCH_ITEMS_SUCCESS } from '../../../constants/actionTypes';
 const fetchingFlagReducer = (state = false, action: IAction): boolean => {
   switch (action.type) {
     case FETCH_ITEMS_REQUEST:
