+++ onboarding-task/src/reducers/rootReducer.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
-import { itemsReducer } from './ListItemReducers/itemsReducer';
-import { itemIdsReducer } from './ListItemReducers/itemIdsReducer';
-import { fetchingFlagReducer } from './fetchingFlagReducer';
-import { errorReducer } from './errorReducer';
-import { postingFlagReducer } from './postingFlagReducer';
-import { errorIdsReducer } from './errorIdsReducer';
+import { itemsReducer } from './ListItemReducers/itemsReducers/itemsReducer';
+import { itemIdsReducer } from './ListItemReducers/itemsReducers/itemIdsReducer';
+import { fetchingFlagReducer } from './ListItemReducers/flagReducers/fetchingFlagReducer';
+import { errorReducer } from './ListItemReducers/errorReducers/errorReducer';
+import { postingFlagReducer } from './ListItemReducers/flagReducers/postingFlagReducer';
+import { errorIdsReducer } from './ListItemReducers/errorReducers/errorIdsReducer';
 
 const rootReducer = combineReducers({
   items: itemsReducer,
