+++ onboarding-task/test/reducerTests/ListItemsReducers/flagReducers/fetchingFlagReducer.test.js
@@ -1,5 +1,5 @@
-import { fetchingFlagReducer } from '../../src/reducers/fetchingFlagReducer.ts';
-import { FETCH_ITEMS_REQUEST, FETCH_ITEMS_SUCCESS, FETCH_ITEMS_FAILURE } from '../../src/constants/actionTypes.ts';
+import { fetchingFlagReducer } from '../../../../src/reducers/ListItemReducers/flagReducers/fetchingFlagReducer.ts';
+import { FETCH_ITEMS_REQUEST, FETCH_ITEMS_SUCCESS, FETCH_ITEMS_FAILURE } from '../../../../src/constants/actionTypes.ts';
 
 describe('Correctly changes the fetch items flag', () => {
   it('changes the fetchingFlag when requesting items to true', () => {
